# api/openapi.yaml
openapi: 3.1.0
info:
  title: Somra API
  version: 0.1.0-dev
  description: |
    Somra — all-in-one, self-hosted home media server.

    This document is the **design-first** source of truth for the Somra HTTP API.
    Frontend TypeScript types are generated from this file (see
    `scripts/gen-openapi-types.sh` and `make openapi-types`). The Go backend
    implements the contract defined here; any divergence is a bug in the
    implementation, not in this specification.

    Errors follow a `ProblemDetails`-style envelope: every error response uses
    the `ErrorResponse` schema, which is a superset compatible with RFC 7807
    "Problem Details for HTTP APIs" (the `messageKey` field carries an
    i18n key resolvable by the client; `message` carries the
    already-localized human text). See `plan/i18n-localization.md` for the
    binding i18n rules.
  license:
    name: AGPL-3.0-or-later
    identifier: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
  contact:
    name: Somra
    url: https://github.com/somralab/somra-media
servers:
  - url: /api/v1
    description: Default API root (relative to the Somra server host)
security:
  - {}
...
paths:
...
  /system/detect:
    get:
      tags: [system]
      summary: Detect system hardware and validate storage paths
      operationId: detectSystem
      parameters:
        - name: paths
          in: query
          description: Comma-separated paths to validate (optional; defaults to configured data/cache dirs)
          schema:
            type: string
      responses:
        "200":
          description: System profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemProfile"
...
  /onboarding/step:
    post:
      tags: [onboarding]
      summary: Advance or update an onboarding step
      operationId: advanceOnboardingStep
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingStepRequest"
      responses:
        "200":
          description: Updated onboarding state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingState"
  /onboarding/complete:
    post:
      tags: [onboarding]
      summary: Mark onboarding as completed
      operationId: completeOnboarding
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Onboarding completed
  /settings:
    get:
      tags: [settings]
      summary: Get all settings grouped by category
      operationId: getSettings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Settings snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsSnapshot"
  /settings/{category}:
    patch:
      tags: [settings]
      summary: Update settings in a category
      operationId: patchSettingsCategory
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
            enum: [general, library, playback, subtitles, streaming]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsCategoryPatch"
      responses:
        "200":
          description: Updated category settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsCategory"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Short-lived JWT access token issued by the auth module. The matching
        revocable refresh token is delivered out-of-band (HTTP-only cookie).
        Not applied to any operation in this revision; reserved for the auth
        endpoints introduced in Sprint 03.
  schemas:
...
    SystemProfile:
      type: object
      required: [cpuCores, memoryBytes, gpuPresent, paths]
      properties:
        cpuCores:
          type: integer
        memoryBytes:
          type: integer
          format: int64
        gpuPresent:
          type: boolean
        paths:
          type: array
          items:
            $ref: "#/components/schemas/PathValidation"
    PathValidation:
      type: object
      required: [path, readable, writable]
      properties:
        path:
          type: string
        readable:
          type: boolean
        writable:
          type: boolean
    SmartDefaults:
      type: object
      required: [maxConcurrentTranscodes, scanCron]
      properties:
        maxConcurrentTranscodes:
          type: integer
        scanCron:
          type: string
        defaultLocale:
          type: string
          enum: [en-US, tr-TR]
...
    OnboardingPhase:
      type: string
      enum: [language, admin, library, defaults, scan, complete]
    OnboardingState:
      type: object
      required: [phase, completed, setupRequired]
      properties:
        phase:
          $ref: "#/components/schemas/OnboardingPhase"
        completed:
          type: boolean
        setupRequired:
          type: boolean
        smartDefaults:
          $ref: "#/components/schemas/SmartDefaults"
    OnboardingStepRequest:
      type: object
      required: [phase]
      properties:
        phase:
          $ref: "#/components/schemas/OnboardingPhase"
        locale:
          type: string
          enum: [en-US, tr-TR]
        applyDefaults:
          type: boolean
        libraryId:
          type: integer
          format: int64
    SettingsSnapshot:
      type: object
      properties:
        general:
          $ref: "#/components/schemas/SettingsCategory"
        library:
          $ref: "#/components/schemas/SettingsCategory"
        playback:
          $ref: "#/components/schemas/SettingsCategory"
        subtitles:
          $ref: "#/components/schemas/SettingsCategory"
        streaming:
          $ref: "#/components/schemas/SettingsCategory"
    SettingsCategory:
      type: object
      additionalProperties: true
    SettingsCategoryPatch:
      type: object
      additionalProperties: true
...